requirements.txt
boto==2.5.2
cmsmenu-redirect==0.1
cmsplugin_simple_markdown
django-admin-tools==0.4.0
django-apptemplates==0.0.1
django-bootstrap-pagination==0.1.10
django-classy-tags==0.3.4.1
django-cms==2.3.4

# this is to make it work with sekizai... when next release after 1.2 ships use that instead
git+https://github.com/jezdez/django_compressor.git@1a7f1ff7cbd2605dc5697a9f2df649d97ff3a669#egg=django-compressor
django-data-exports==0.2
django-debug-toolbar==0.9.4
django-extensions==0.7
django-floppyforms==1.0
django-fsm==1.3.0
django-inspect-model==0.4
django-jsonfield==0.8.10
django-kronos==0.3
django-mptt==0.5.2
django-nose==1.1
django-registration==0.8
django-reversion==1.6.5
django-sekizai==0.6.1
django-selectable==0.2
django-storages==1.1.5
django-tables==0.2
django-tastypie==0.9.11
Django==1.4.3
django_admin_bootstrapped==0.1.1
docutils==0.9.1
envoy==0.0.2
html5lib==0.95
Markdown==2.2.1
more-itertools==2.0
nose==1.1.2
paramiko==1.7.7.2
phonenumbers==5.0b2
pillow==1.7.8
python-dateutil==2.1
python-magic==0.4.3
South==0.7.6
Sphinx==1.1.3
statlib==1.1
twilio==3.3.9
pyparsing==1.5.6
parsedatetime==0.8.7
psycopg2==2.4.1
dj-database-url==0.2.1
django-smtp-ssl==1.0
django-secure==0.1.2
